moved schema type defs to seperate file
