+++ src/utils/load_testing/routeSchema.ts
@@ -1,0 +1,6 @@
+export type RouterSchema = Record<string, ProcedureInfo>;
+export type ProcedureInfo = {
+  path: string;
+  input: string;
+  type: "query" | "mutation" | "subscription";
+};
